refactor: extract UI/logic violations into dedicated hooks
- MazeClient: remove direct useMazeStore access; expose move via useMazeGame
- memoryStoreTypes: remove duplicate MemoryPhase definition; re-export from types/memory
- VocabularyOfTheDay, DailyChallenge, RecentlyPlayedRow, ContinueBanner, OnboardingModal:
  extract localStorage/sessionStorage, TTS, and data-fetching logic into
  hooks/shared/marketing/ hooks
- GameCard: extract getMasteryStars useEffect into hooks/shared/progress/useMasteryStars
